requirements_production.txt
Flask==3.0.0
gunicorn==21.2.0
praw==7.7.1
gspread==5.12.0
google-auth==2.27.0
openpyxl==3.1.2
pandas==2.2.0
requests==2.31.0
python-dotenv==1.0.0